Crate whitenoise_runtime

The Whitenoise rust runtime is an execution engine for evaluating differentially private analyses.

The runtime contains implementations of basic data transformations and aggregations, statistics, and privatizing mechanisms. These functions are combined in the Whitenoise validator to create more complex differentially private analyses.

Modules

components
Component trait implementations

ffi

proto

utilities

Functions

execute_graph
Given a description of computation, and some computed values, execute the computation and return computed values

release
Evaluate an analysis and release the differentially private results.

Type Definitions

NodeArguments