Function whitenoise_runtime::components::impute::impute_float_uniform

pub fn impute_float_uniform(
    data: &ArrayD<f64>, 
    min: &ArrayD<f64>, 
    max: &ArrayD<f64>
) -> Result<ArrayD<f64>>

Returns data with imputed values in place of f64::NAN. Values are imputed from a uniform distribution.

Arguments

data - Data for which you would like to impute the NAN values.

min - Lower bound on imputation range for each column.

max - Upper bound on imputation range for each column.

Return

Data with NAN values replaced with imputed values.

Example

use ndarray::prelude::*;
use whitenoise_runtime::components::impute::impute_float_uniform;
use core::f64::NAN;

let data: ArrayD<f64> = arr2(&[ [1., NAN, 3., NAN], [2., 2., NAN, NAN] ]).into_dyn();
let min: ArrayD<f64> = arr1(&[0., 2., 3., 4.]).into_dyn();
let max: ArrayD<f64> = arr1(&[10., 2., 5., 5.]).into_dyn();
let imputed = impute_float_uniform(&data, &min, &max);