Module whitenoise_runtime::components::resize

Functions

create_sampling_indices
Accepts size of set (n) and size of desired subset(k) and returns a uniformly drawn set of indices from [1, ..., n] of size k.

resize_categorical
Resizes categorical data based on estimate of n and true size of data.

resize_float
Resizes data (made up exclusively of f64) based on estimate of n and true size of data.

resize_integer
Resizes data (made up exclusively of i64) based on estimate of n and true size of data.