Function whitenoise_runtime::components::resize::resize_float

pub fn resize_float(
    data: &ArrayD<f64>, 
    n: &i64, 
    distribution: &String, 
    min: &ArrayD<f64>, 
    max: &ArrayD<f64>, 
    shift: &Option<&ArrayD<f64>>, 
    scale: &Option<&ArrayD<f64>>
) -> Result<ArrayD<f64>>

Resizes data (made up exclusively of f64) based on estimate of n and true size of data.

Notice that some arguments are denoted with Option<> -- this is because not every distribution used for imputation (if necessary) uses every argument (e.g. Uniform does not use shift or scale).

NOTE: If more distributions are added here, their corresponding arguments must be added as inputs.

Arguments

data - The data to be resized

n - An estimate of the size of the data -- this could be the guess of the user, or the result of a DP release

distribution - The distribution to be used when imputing records

min - A lower bound on data elements

max - An upper bound on data elements

shift - The shift (expectation) argument for the Gaussian distribution

scale - The scale (standard deviation) argument for the Gaussian distribution

Return

A resized version of data consistent with the provided n