Function whitenoise_runtime::execute_graph

pub fn execute_graph(
    analysis: &Analysis, 
    release: &Release, 
    filter_level: &FilterLevel
) -> Result<(Release, Vec<Error>)>

Given a description of computation, and some computed values, execute the computation and return computed values

The analysis is a directed acyclic graph.

vertices are components (a unit of computation)

edges are arguments

When a component is executed, the output of the node is stored in the release When the graph completes execution, the release is filtered and returned

Arguments

analysis - a computational graph and definition of privacy, in prost protobuf format

release - a collection of precomputed values for components in the graph

filter_level - configure the amount of information included in the return

Return

a collection of computed values for components in the graph