Trait whitenoise_runtime::components::Evaluable

pub trait Evaluable {
    fn evaluate(
        &self, 
        privacy_definition: &Option<PrivacyDefinition>, 
        arguments: NodeArguments
    ) -> Result<ReleaseNode>;
}

Evaluable component trait

Evaluable structs represent an abstract computation.

Required methods

fn evaluate(
&self,
privacy_definition: &Option<PrivacyDefinition>,
arguments: NodeArguments
) -> Result<ReleaseNode>

The concrete implementation of the abstract computation that the struct represents.

Arguments

privacy_definition - the definition of privacy under which the computation takes place

arguments - a hashmap, where the String keys are the names of arguments, and the Value values are the data inputs

Returns

The concrete value corresponding to the abstract computation that the struct represents

Loading content...